Accept calendar event
In order to accept the event, we need to build an `Event` value with the
data that should be patched into the original. That value is then passed
in the `patch()` call to update the event.

The email address of the current user is required. I couldn't figure out
how best to get it, so I decided to keep the event `get()` call, and
grab the authenticated user's email address from there. This means it
requires two requests to accept the event, which is not ideal. Not sure
if there's a better way to get that information.