event_id_from_base64(): Assume `event_id` is base64-encoded
A base64-encoded string I tested with didn't pass the regex test, and
was treated as an event ID even though it was an `eid`.

Instead of testing whether the input is a base64-encoded string with
regex, assume it's a base64-encoded string and try to decode it. If that
fails, assume it's a regular event ID.