Fix "token.json" not found error
Because I was using `xdg_dirs.find_data_file()` to get the path to
"token.json", an error was triggered when the file didn't exist yet. It
doesn't need to exist because the program is supposed to create it, so
we can't use `find_data_file()`.

Remove the `local_data_file()` function, and instead call the XDG
functions directly where they're needed since in one case we need to
find an existing file (and error with a specific message if it doesn't
exist), and in another, we need a path to a file to create or read.